backend/requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djongo==1.3.6
pymongo==3.12.3
python-dotenv==1.0.1